refactor: rename every camelCase file to kebab-case
