+++ src/config/social-medias.ts
