+++ src/core/types/paginated-response.ts
@@ -1,0 +1,5 @@
+export type PaginatedResponse = {
+	totalCount: number;
+	page: number;
+	perPage: number;
+};
