--- src/core/types/paginatedResponse.ts
@@ -1,5 +1,0 @@
-export type PaginatedResponse = {
-	totalCount: number;
-	page: number;
-	perPage: number;
-};
