+++ src/core/types/server-response-data.ts
