+++ src/routes/admin/(authorized)/projetos/delete-project-button.svelte
