+++ src/routes/admin/(authorized)/projetos/handlers.ts
@@ -1,7 +1,7 @@
 import type { Project } from "$crate/core/entities/project";
-import { MakeServerResponseData } from "$crate/core/helpers/serverActionResponse";
-import type { PaginatedResponse } from "$crate/core/types/paginatedResponse";
-import type { ServerResponseData } from "$crate/core/types/serverResponseData";
+import { MakeServerResponseData } from "$crate/core/helpers/server-action-response";
+import type { PaginatedResponse } from "$crate/core/types/paginated-response";
+import type { ServerResponseData } from "$crate/core/types/server-response-data";
 import { env } from "$env/dynamic/private";
 import { fail, type ActionFailure, type RequestEvent, type ServerLoadEvent } from "@sveltejs/kit";
